fixing people types

+++ InvenShopfy.Core/Models/People/Biller.cs
@@ -15,7 +15,7 @@ public class Biller
     public string Address { get; set; } = string.Empty;
     public string Country { get; set; } = string.Empty;
     public string ZipCode { get => _zipCode; set => _zipCode = _zipCodeFormatter.FormatZipCode(value); }
-    public double BillerCode { get; set; } 
+    public long BillerCode { get; set; } 
     public string UserId { get; set; } = string.Empty;
     
     public long WarehouseId { get; set; }
